+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fetch token prices based on contract address, not symbol
 - Fix bug that prevents setting language locale in settings.
 - Show checksum addresses throughout the UI
+- Allow zero gas price on testnet.
 
 ## 4.5.5 Fri Apr 06 2018